requirements.txt
werkzeug
jinja2
python-slugify
